refactor: split import handler into normalize, s3_ops, persist modules
Extract the monolithic process_import_job function into focused modules:
- normalize.rs: pure functions for severity/subject_kind/coordinate normalization
- s3_ops.rs: S3 download, verify, extract and upload operations
- persist.rs: all DB persistence (artifacts, checks, snapshot, diff, completion)
- mod.rs: orchestrator with validate_payload, process_import_job, handle_import_failure

No behavioral changes. Transaction boundary preserved in process_import_job.

Closes #100
