+++ crates/worker/src/handlers/import/mod.rs
@@ -1,11 +1,10 @@
-use boardflow_artifact::{
-    ArtifactError, download_bundle, extract_bundle, upload_artifact, verify_sha256,
-};
-use boardflow_db::queries::{
-    artifact, artifact_bundle, board_project, board_run, diff, github_job, run_check,
-    run_check_finding, snapshot,
-};
-use boardflow_domain::models::github_job::{GithubJob, GithubJobType};
+mod normalize;
+mod persist;
+mod s3_ops;
+
+use boardflow_artifact::ArtifactError;
+use boardflow_db::queries::{artifact_bundle, board_run, github_job};
+use boardflow_domain::models::github_job::GithubJob;
 use boardflow_jobs::{BASE_BACKOFF_SECS, MAX_ATTEMPTS};
 use serde::Deserialize;
 use sqlx::PgPool;
@@ -39,12 +38,18 @@ pub async fn handle(
     }
 }
 
-async fn process_import_job(
+async fn validate_payload(
     pool: &PgPool,
-    s3_client: &aws_sdk_s3::Client,
-    config: &WorkerConfig,
     job: &GithubJob,
-) -> Result<(), ArtifactError> {
+) -> Result<
+    (
+        ImportPayload,
+        Uuid,
+        Uuid,
+        boardflow_domain::models::artifact_bundle::ArtifactBundle,
+    ),
+    ArtifactError,
+> {
     let payload: ImportPayload = serde_json::from_value(job.payload_json.clone())
         .map_err(|e| ArtifactError::Manifest(format!("invalid job payload: {e}")))?;
 
@@ -66,431 +71,51 @@ async fn process_import_job(
         .await
         .map_err(|e| ArtifactError::S3(e.to_string()))?;
 
-    // === Pre-transaction: S3 operations ===
+    Ok((payload, board_run_id, board_project_id, bundle))
+}
 
-    // Step 1: Download bundle from S3
-    tracing::info!(key = %payload.staging_object_key, "Downloading bundle from S3");
-    let data = download_bundle(
+async fn process_import_job(
+    pool: &PgPool,
+    s3_client: &aws_sdk_s3::Client,
+    config: &WorkerConfig,
+    job: &GithubJob,
+) -> Result<(), ArtifactError> {
+    // Phase 1: Validate payload and mark bundle as importing
+    let (payload, board_run_id, board_project_id, bundle) = validate_payload(pool, job).await?;
+
+    // Phase 2: S3 operations (pre-transaction)
+    let data = s3_ops::download_and_verify(
         s3_client,
-        &config.s3.staging_bucket,
+        config,
         &payload.staging_object_key,
+        &payload.bundle_sha256,
     )
     .await?;
+    let (manifest, uploaded) =
+        s3_ops::extract_and_upload(s3_client, config, &data, board_run_id).await?;
 
-    // Step 2: Verify SHA256
-    verify_sha256(&data, &payload.bundle_sha256)?;
-
-    // Step 3: Extract bundle
-    tracing::info!("Extracting bundle");
-    let (manifest, extracted_artifacts) = extract_bundle(&data)?;
-
-    // Step 4: Upload artifacts to final bucket (before transaction)
-    tracing::info!(
-        count = extracted_artifacts.len(),
-        "Uploading artifacts to final bucket"
-    );
-
-    // Prepare upload results for use inside transaction
-    struct UploadedArtifact {
-        storage_key: String,
-        sha256: String,
-        size: i64,
-        manifest_idx: usize,
-    }
-    let mut uploaded: Vec<UploadedArtifact> = Vec::new();
-
-    for (idx, manifest_entry) in manifest.artifacts.iter().enumerate() {
-        if manifest_entry.status != "available" {
-            continue;
-        }
-        let storage_key = format!(
-            "artifacts/{board_run_id}/{}/{}",
-            manifest_entry.r#type, manifest_entry.filename
-        );
-
-        if let Some(extracted) = extracted_artifacts
-            .iter()
-            .find(|a| manifest_entry.source_path.as_deref() == Some(&a.path))
-        {
-            upload_artifact(
-                s3_client,
-                &config.s3.final_bucket,
-                &storage_key,
-                extracted.data.clone(),
-                &manifest_entry.content_type,
-            )
-            .await?;
-
-            uploaded.push(UploadedArtifact {
-                storage_key,
-                sha256: extracted.sha256.clone(),
-                size: extracted.data.len() as i64,
-                manifest_idx: idx,
-            });
-        }
-    }
-
-    // === Transaction: all DB writes ===
+    // Phase 3: Transaction — all DB writes
     let mut tx = pool
         .begin()
         .await
         .map_err(|e| ArtifactError::S3(e.to_string()))?;
 
-    // Insert artifacts (available ones that were uploaded)
-    for up in &uploaded {
-        let manifest_entry = &manifest.artifacts[up.manifest_idx];
-        artifact::insert(
-            &mut *tx,
-            Uuid::now_v7(),
-            board_run_id,
-            manifest_entry.r#type,
-            "available",
-            Some(&manifest_entry.filename),
-            manifest_entry.source_path.as_deref(),
-            manifest_entry.logical_name.as_deref(),
-            Some(&manifest_entry.content_type),
-            Some(&up.storage_key),
-            Some(&up.sha256),
-            Some(up.size),
-            None,
-            Some(bundle.id),
-        )
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-    }
-
-    // Insert non-available artifacts
-    for manifest_entry in &manifest.artifacts {
-        if manifest_entry.status == "available" {
-            continue;
-        }
-        artifact::insert(
-            &mut *tx,
-            Uuid::now_v7(),
-            board_run_id,
-            manifest_entry.r#type,
-            &manifest_entry.status,
-            Some(&manifest_entry.filename),
-            manifest_entry.source_path.as_deref(),
-            manifest_entry.logical_name.as_deref(),
-            Some(&manifest_entry.content_type),
-            None,
-            None,
-            None,
-            manifest_entry.status_reason.as_deref(),
-            Some(bundle.id),
-        )
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-    }
-
-    // Step 5: Save run checks (ERC/DRC)
-    let mut erc_status: Option<&str> = None;
-    let mut erc_errors = 0i32;
-    let mut erc_warnings = 0i32;
-    let mut drc_status: Option<&str> = None;
-    let mut drc_errors = 0i32;
-    let mut drc_warnings = 0i32;
-
-    for check in &manifest.checks {
-        let check_id = Uuid::now_v7();
-        run_check::insert(
-            &mut *tx,
-            check_id,
-            board_run_id,
-            &check.kind,
-            &check.status,
-            check.error_count,
-            check.warning_count,
-            check.notice_count,
-            check.tool_name.as_deref(),
-            check.tool_version.as_deref(),
-            check.raw_summary.as_ref(),
-        )
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-        // Insert findings for this check
-        for (idx, raw_finding) in check.findings.iter().enumerate() {
-            match serde_json::from_value::<boardflow_artifact::ManifestFinding>(raw_finding.clone())
-            {
-                Ok(finding) => {
-                    // Normalize severity to pass DB CHECK constraint
-                    let severity = match finding.severity.as_str() {
-                        "error" | "warning" | "notice" => finding.severity.as_str(),
-                        _ => "notice",
-                    };
-
-                    // Normalize subject_kind to pass DB CHECK constraint
-                    let subject_kind = finding.subject_kind.as_deref().and_then(|sk| match sk {
-                        "schematic" | "pcb" | "net" | "footprint" | "symbol" => Some(sk),
-                        _ => None,
-                    });
-
-                    let x_um = finding
-                        .pos_mm
-                        .as_ref()
-                        .map(|p| (p.x * 1000.0).round() as i32);
-                    let y_um = finding
-                        .pos_mm
-                        .as_ref()
-                        .map(|p| (p.y * 1000.0).round() as i32);
-
-                    if let Err(e) = run_check_finding::insert(
-                        &mut *tx,
-                        Uuid::now_v7(),
-                        check_id,
-                        severity,
-                        Some(&finding.rule_code),
-                        Some(&finding.title),
-                        finding.message.as_deref(),
-                        subject_kind,
-                        finding.subject_ref.as_deref(),
-                        finding.sheet_path.as_deref(),
-                        finding.pcb_layer.as_deref(),
-                        x_um,
-                        y_um,
-                        None,
-                        finding.raw.as_ref(),
-                        idx as i32,
-                    )
-                    .await
-                    {
-                        tracing::error!(
-                            check_id = %check_id,
-                            sort_index = idx,
-                            error = %e,
-                            "Failed to insert finding after normalization, skipping"
-                        );
-                    }
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        check_id = %check_id,
-                        sort_index = idx,
-                        error = %e,
-                        "Failed to parse finding, storing raw payload"
-                    );
-                    if let Err(insert_err) = run_check_finding::insert(
-                        &mut *tx,
-                        Uuid::now_v7(),
-                        check_id,
-                        "notice",
-                        None,
-                        None,
-                        None,
-                        None,
-                        None,
-                        None,
-                        None,
-                        None,
-                        None,
-                        None,
-                        Some(raw_finding),
-                        idx as i32,
-                    )
-                    .await
-                    {
-                        tracing::error!(
-                            check_id = %check_id,
-                            sort_index = idx,
-                            error = %insert_err,
-                            "Failed to insert raw finding fallback, skipping"
-                        );
-                    }
-                }
-            }
-        }
-
-        match check.kind.as_str() {
-            "erc" => {
-                erc_status = Some(if check.status == "passed" {
-                    "passed"
-                } else if check.status == "failed" {
-                    "failed"
-                } else {
-                    "skipped"
-                });
-                erc_errors = check.error_count;
-                erc_warnings = check.warning_count;
-            }
-            "drc" => {
-                drc_status = Some(if check.status == "passed" {
-                    "passed"
-                } else if check.status == "failed" {
-                    "failed"
-                } else {
-                    "skipped"
-                });
-                drc_errors = check.error_count;
-                drc_warnings = check.warning_count;
-            }
-            _ => {}
-        }
-    }
-
-    // Step 6: Create snapshot
-    let file_hashes_json = serde_json::to_value(&manifest.files)
-        .map_err(|e| ArtifactError::Manifest(e.to_string()))?;
-    snapshot::insert(
-        &mut *tx,
-        Uuid::now_v7(),
+    persist::persist_artifacts(&mut tx, board_run_id, bundle.id, &manifest, &uploaded).await?;
+    let check_summary =
+        persist::persist_checks_and_findings(&mut tx, board_run_id, &manifest).await?;
+    persist::persist_snapshot_and_diff(&mut tx, board_project_id, board_run_id, &manifest).await?;
+    persist::complete_run(
+        &mut tx,
         board_project_id,
         board_run_id,
+        bundle.id,
+        job.id,
         &manifest.tree_hash,
-        &manifest.commit_sha,
-        &file_hashes_json,
+        &check_summary,
     )
-    .await
-    .map_err(|e| ArtifactError::S3(e.to_string()))?;
+    .await?;
+    persist::enqueue_follow_up_jobs(&mut tx, job, board_project_id, board_run_id).await?;
 
-    // Step 7: Save diff metadata
-    if let Some(ref dm) = manifest.diff_metadata {
-        diff::insert_diff_metadata(
-            &mut *tx,
-            Uuid::now_v7(),
-            board_run_id,
-            dm.file_hashes.as_ref(),
-            dm.bom_summary.as_ref(),
-            dm.checks_summary.as_ref(),
-            dm.artifacts_summary.as_ref(),
-            dm.previews.as_ref(),
-        )
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-    }
-
-    // Step 8: Resolve baseline and create diff record
-    let bp = board_project::find_by_id(&mut *tx, board_project_id)
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-    let base_run_id = bp.and_then(|p| p.latest_completed_run_id);
-    let diff_status = if base_run_id.is_some() {
-        "ready"
-    } else {
-        "no_baseline"
-    };
-
-    // Compute diff summary from file_hashes if we have a baseline
-    let summary_json = if base_run_id.is_some() {
-        let file_count = manifest.files.len();
-        Some(serde_json::json!({
-            "total_files": file_count,
-            "status": "computed"
-        }))
-    } else {
-        None
-    };
-
-    diff::insert_diff(
-        &mut *tx,
-        Uuid::now_v7(),
-        board_run_id,
-        base_run_id,
-        diff_status,
-        summary_json.as_ref(),
-    )
-    .await
-    .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    // Step 9: Mark board_run as completed
-    board_run::mark_completed(
-        &mut *tx,
-        board_run_id,
-        erc_status,
-        erc_errors,
-        erc_warnings,
-        drc_status,
-        drc_errors,
-        drc_warnings,
-    )
-    .await
-    .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    // Step 10: Update board_project latest_completed_run_id
-    board_project::update_latest_completed_run(
-        &mut *tx,
-        board_project_id,
-        board_run_id,
-        &manifest.tree_hash,
-    )
-    .await
-    .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    // Step 11: Mark bundle as completed
-    artifact_bundle::mark_completed(&mut *tx, bundle.id)
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    // Step 12: Mark job as completed
-    github_job::mark_completed(&mut *tx, job.id)
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    // Step 13: Enqueue follow-up jobs
-
-    // 1. Issue sync: if board_project has no issue_number, enqueue create_issue
-    let bp_for_jobs = board_project::find_by_id(&mut *tx, board_project_id)
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    if bp_for_jobs
-        .as_ref()
-        .and_then(|bp| bp.issue_number)
-        .is_none()
-    {
-        let _ = github_job::enqueue(
-            &mut *tx,
-            Uuid::now_v7(),
-            job.installation_id,
-            job.repository_id,
-            Some(board_project_id),
-            Some(board_run_id),
-            GithubJobType::CreateIssue,
-            &serde_json::json!({}),
-        )
-        .await
-        .map_err(|e| ArtifactError::S3(e.to_string()))?;
-    }
-
-    // 2. Dashboard comment: create or update based on board_project.dashboard_comment_id
-    let dashboard_job_type = if bp_for_jobs
-        .as_ref()
-        .and_then(|bp| bp.dashboard_comment_id)
-        .is_some()
-    {
-        GithubJobType::UpdateDashboardComment
-    } else {
-        GithubJobType::CreateDashboardComment
-    };
-    let _ = github_job::enqueue(
-        &mut *tx,
-        Uuid::now_v7(),
-        job.installation_id,
-        job.repository_id,
-        Some(board_project_id),
-        Some(board_run_id),
-        dashboard_job_type,
-        &serde_json::json!({}),
-    )
-    .await
-    .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    // 3. Run result comment
-    let _ = github_job::enqueue(
-        &mut *tx,
-        Uuid::now_v7(),
-        job.installation_id,
-        job.repository_id,
-        Some(board_project_id),
-        Some(board_run_id),
-        GithubJobType::CreateRunResultComment,
-        &serde_json::json!({}),
-    )
-    .await
-    .map_err(|e| ArtifactError::S3(e.to_string()))?;
-
-    // Commit transaction
     tx.commit()
         .await
         .map_err(|e| ArtifactError::S3(e.to_string()))?;
